Update: implement update name
