+++ src/model/types.rs
@@ -1,5 +1,6 @@
+use axum::response::IntoResponse;
+use hyper::{Body, Response, StatusCode};
 use serde::{Deserialize, Serialize};
-// use validator::Validate;
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct UserId {
@@ -12,5 +13,29 @@ pub enum SomeError {
     NotExistError,
     InvalidPasswordError,
     SameNameError,
-    InvalidStructure
+    InvalidStructure,
+}
+
+impl IntoResponse for SomeError {
+    type Body = Body;
+    type BodyError = <Self::Body as axum::body::HttpBody>::Error;
+    fn into_response(self) -> Response<Self::Body> {
+        let body = match self {
+            SomeError::ValidationError => Body::from("invalid validation"),
+            SomeError::NotExistError => Body::from("user not found"),
+            SomeError::SameNameError => Body::from("the name is alreasy used"),
+            SomeError::InvalidPasswordError => Body::from("user not found"),
+            SomeError::InvalidStructure => Body::from("invalid structure"),
+        };
+
+        let status = match self {
+            SomeError::ValidationError => StatusCode::UNPROCESSABLE_ENTITY,
+            SomeError::NotExistError => StatusCode::NOT_FOUND,
+            SomeError::SameNameError => StatusCode::BAD_REQUEST,
+            SomeError::InvalidPasswordError => StatusCode::NOT_FOUND,
+            SomeError::InvalidStructure => StatusCode::BAD_REQUEST,
+        };
+
+        Response::builder().status(status).body(body).unwrap()
+    }
 }
